# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AgentFlow",
    "description": "Multi-agent orchestration system",
    "version": "0.1.0"
  },
  "paths": {
    "/api/runs": {
      "get": {
        "summary": "List Runs",
        "operationId": "list_runs_api_runs_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunListResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Start Run",
        "operationId": "start_run_api_runs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/stream": {
      "get": {
        "summary": "Stream Run",
        "operationId": "stream_run_api_runs__run_id__stream_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}": {
      "get": {
        "summary": "Get Run",
        "operationId": "get_run_api_runs__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunInfo"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/events": {
      "get": {
        "summary": "Get Run Events",
        "operationId": "get_run_events_api_runs__run_id__events_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunEventsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/results": {
      "get": {
        "summary": "Get Run Results",
        "operationId": "get_run_results_api_runs__run_id__results_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResultsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/report": {
      "get": {
        "summary": "Get Run Report",
        "operationId": "get_run_report_api_runs__run_id__report_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunReportResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/artifacts": {
      "get": {
        "summary": "Get Run Artifacts",
        "operationId": "get_run_artifacts_api_runs__run_id__artifacts_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunArtifactsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/artifacts/{artifact_id}": {
      "get": {
        "summary": "Get Run Artifact Content",
        "operationId": "get_run_artifact_content_api_runs__run_id__artifacts__artifact_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "artifact_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Artifact Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunArtifactContentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "summary": "Health",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Health Health Get"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AgentOutput": {
        "properties": {
          "structured": {
            "additionalProperties": true,
            "type": "object",
            "title": "Structured"
          },
          "text": {
            "type": "string",
            "title": "Text",
            "default": ""
          }
        },
        "type": "object",
        "title": "AgentOutput"
      },
      "AgentStatus": {
        "type": "string",
        "enum": [
          "success",
          "partial",
          "failed"
        ],
        "title": "AgentStatus"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "RunArtifact": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "path": {
            "type": "string",
            "title": "Path"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "path"
        ],
        "title": "RunArtifact"
      },
      "RunArtifactContentResponse": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "artifact_id": {
            "type": "string",
            "title": "Artifact Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "path": {
            "type": "string",
            "title": "Path"
          },
          "content": {
            "type": "string",
            "title": "Content"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "artifact_id",
          "name",
          "path",
          "content"
        ],
        "title": "RunArtifactContentResponse"
      },
      "RunArtifactsResponse": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "artifacts": {
            "items": {
              "$ref": "#/components/schemas/RunArtifact"
            },
            "type": "array",
            "title": "Artifacts"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "artifacts"
        ],
        "title": "RunArtifactsResponse"
      },
      "RunEventsResponse": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/SSEEvent"
            },
            "type": "array",
            "title": "Events"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "events"
        ],
        "title": "RunEventsResponse"
      },
      "RunInfo": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "has_events": {
            "type": "boolean",
            "title": "Has Events"
          },
          "has_results": {
            "type": "boolean",
            "title": "Has Results"
          },
          "has_report": {
            "type": "boolean",
            "title": "Has Report"
          },
          "has_artifacts": {
            "type": "boolean",
            "title": "Has Artifacts",
            "default": false
          },
          "task": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Task"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "has_events",
          "has_results",
          "has_report"
        ],
        "title": "RunInfo"
      },
      "RunListResponse": {
        "properties": {
          "runs": {
            "items": {
              "$ref": "#/components/schemas/RunInfo"
            },
            "type": "array",
            "title": "Runs"
          }
        },
        "type": "object",
        "required": [
          "runs"
        ],
        "title": "RunListResponse"
      },
      "RunReportResponse": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "report": {
            "type": "string",
            "title": "Report"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "report"
        ],
        "title": "RunReportResponse"
      },
      "RunRequest": {
        "properties": {
          "task": {
            "type": "string",
            "title": "Task"
          },
          "context": {
            "additionalProperties": true,
            "type": "object",
            "title": "Context"
          },
          "budget_usd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Budget Usd"
          }
        },
        "type": "object",
        "required": [
          "task"
        ],
        "title": "RunRequest"
      },
      "RunResponse": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "started"
          }
        },
        "type": "object",
        "required": [
          "run_id"
        ],
        "title": "RunResponse"
      },
      "RunResultsResponse": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/SubtaskResult"
            },
            "type": "array",
            "title": "Results"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "results"
        ],
        "title": "RunResultsResponse"
      },
      "SSEEvent": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "seq": {
            "type": "integer",
            "title": "Seq"
          },
          "ts": {
            "type": "integer",
            "title": "Ts"
          },
          "type": {
            "$ref": "#/components/schemas/SSEEventType"
          },
          "agent_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agent Id"
          },
          "payload": {
            "$ref": "#/components/schemas/SSEPayload"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "seq",
          "type"
        ],
        "title": "SSEEvent"
      },
      "SSEEventType": {
        "type": "string",
        "enum": [
          "run:started",
          "plan:created",
          "task:dispatched",
          "agent:progress",
          "agent:query",
          "task:complete",
          "task:partial",
          "task:failed",
          "task:continuing",
          "run:complete",
          "run:error",
          "run:budget_exceeded"
        ],
        "title": "SSEEventType"
      },
      "SSEPayload": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "default": ""
          },
          "partial": {
            "title": "Partial"
          },
          "data": {
            "title": "Data"
          }
        },
        "type": "object",
        "title": "SSEPayload"
      },
      "SubtaskResult": {
        "properties": {
          "task_id": {
            "type": "string",
            "title": "Task Id"
          },
          "agent_id": {
            "type": "string",
            "title": "Agent Id"
          },
          "status": {
            "$ref": "#/components/schemas/AgentStatus"
          },
          "output": {
            "$ref": "#/components/schemas/AgentOutput"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "input_tokens": {
            "type": "integer",
            "title": "Input Tokens",
            "default": 0
          },
          "output_tokens": {
            "type": "integer",
            "title": "Output Tokens",
            "default": 0
          },
          "cache_creation_tokens": {
            "type": "integer",
            "title": "Cache Creation Tokens",
            "default": 0
          },
          "cache_read_tokens": {
            "type": "integer",
            "title": "Cache Read Tokens",
            "default": 0
          },
          "tokens_used": {
            "type": "integer",
            "title": "Tokens Used",
            "default": 0
          },
          "cost_usd": {
            "type": "number",
            "title": "Cost Usd",
            "default": 0.0
          },
          "duration_ms": {
            "type": "integer",
            "title": "Duration Ms",
            "default": 0
          },
          "subtask_id": {
            "type": "string",
            "title": "Subtask Id"
          }
        },
        "type": "object",
        "required": [
          "task_id",
          "agent_id",
          "status",
          "subtask_id"
        ],
        "title": "SubtaskResult"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}